add headless build
