+++ graph/oauth2_gtk.h
@@ -1,1 +1,3 @@
+#pragma once
+
 char *webkit_auth_window(char *auth_url);
